Added Api call. Screen Navigation ViewModels

+++ app/src/main/java/app/map/harrypotter/domain/model/Characters.kt
@@ -31,7 +31,7 @@ data class CharactersItem(
 
 data class Wand(
 	val core: String? = null,
-	val length: Int? = null,
+	val length: Double? = null,
 	val wood: String? = null
 )
